Move CSV data read into a function.
Since we will be adding the ability to read
data from xlsx as well as csv files it makes
sense to handle the different data reads in
special functions.  This pulls the existing
capability to read CSV out of main into a
new function.  Also added logic that will
cause the script to fail if an unrecognized
data type is specified in the config file.