Move CSV data write into a function.
Since we will be adding the ability to write
to an Excel (xlsx) file, we will need to
have different output writing techniques
based on the specified output type.  To
prepare for that we will pull the CSV data
writing logic out of main.  Also add logic
to verify that the specified output type
is recognized.